+++ Autodesk.PackageBuilder.Tests/Constants/ProductTypes_Tests.cs
@@ -1,0 +1,19 @@
+﻿using NUnit.Framework;
+
+namespace Autodesk.PackageBuilder.Tests.Constants
+{
+    public class ProductTypes_Tests
+    {
+        [Test]
+        public void Application_Test()
+        {
+            Assert.AreEqual("Application", ProductTypes.Application);
+        }
+
+        [Test]
+        public void Content_Test()
+        {
+            Assert.AreEqual("Content", ProductTypes.Content);
+        }
+    }
+}