+++ Autodesk.PackageBuilder/Model/IPackageSerializable.cs
@@ -1,5 +1,9 @@
 ﻿namespace Autodesk.PackageBuilder.Model
 {
+    /// <summary>
+    /// IPackageSerializable
+    /// </summary>
+    /// <remarks>Interface to SerializeFile with <see cref="SerializeExtensions"/></remarks>
     public interface IPackageSerializable
     {