+++ Autodesk.PackageBuilder/Utils/RevitUtils.cs
@@ -1,9 +1,12 @@
 ﻿namespace Autodesk.PackageBuilder
 {
+    /// <summary>
+    /// RevitUtils
+    /// </summary>
     public static class RevitUtils
     {
         /// <summary>
-        /// Set <see cref="AutodeskProducts.Revit"/> and <see cref="ProductTypes.Application"/>.
+        /// Set AutodeskProduct to <see cref="AutodeskProducts.Revit"/> and ProductType to <see cref="ProductTypes.Application"/>.
         /// </summary>
         /// <param name="applicationPackageBuilder"></param>
         /// <returns></returns>