+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] / 2023-11-24 - 2023-12-06
+### Features
+- `Tests` project.
+- Update `BuilderUtils` to return instance and features (#16).
+### PackageBuilder
+- Enable `DocumentationFile` in package.
+- Update `BuilderUtils` to return instance.
+- Obsolete and internal `AllowLoadingIntoExistingSession`
+- Update `Example` Guid.
+### Tests
+- Test `Constants`
+- Test `BuilderUtils`
+- Test `Builder` create file.
+- Test `RevitAddInsBuilder`
+- Test `PackageContentsBuilder`
+- Add `AssertBuilderUtils`
+- Test `RevitUtils`
+
 ## [1.0.5] / 2021-12-21
 - Update Build Project
 - 2021-12-21